+++ tests/test_unicorn.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from unicorn import matrix_to_bytes, bytes_to_matrix
+from pi.unicorn import matrix_to_bytes, bytes_to_matrix
 from random import randint, randbytes
